rename current excel file to v1, use v2 as default
