+++ .github/workflows/e2e-tests.yml
@@ -7,6 +7,7 @@ on:
       - 'shared/**'
       - 'api/**'
       - 'client/**'
+      - 'data/excel/Carbon-Cost Data Upload.xlsm'
       - '.github/workflows/e2e-tests.yml'
       - '/*'               # include changes in root
       - '!infrastructure/**'  # exclude infra folder
