+++ api/test/integration/import/entity.preprocessor.spec.ts
@@ -44,7 +44,8 @@ describe('Entity Preprocessor', () => {
             price_type: PROJECT_PRICE_TYPE.MARKET_PRICE,
             project_size_ha: 8000,
             project_size_filter: 'Large',
-            abatement_potential: 1044882,
+            project_abatement_potential: 1044882,
+            country_abatement_potential: 1044882,
             total_cost_npv: 4848562,
             total_cost: 7842918,
             capex_npv: 1342839,
